cmake: fix path to FindGLog.cmake
Summary:
The cmake files refer to this package as "FindGLog", but the file name on disk
incorrectly had the 'L' lower cased.  This worked on case-insensitive file
systems, but caused cmake to fail when used on a case-sensitive file system.

This diff fixes the file name to match the name used in the cmake files.  This
is also more consistent with the existing "FindGFlags" and "FindGMock"
packages.

Reviewed By: yfeldblum

Differential Revision: D6710439

fbshipit-source-id: 19f6824ef6793e505b5a0080c0fefe0b145a41c3