Handle max deadlines in Futex
Summary:
[Folly] Handle max deadlines in `Futex`.

Some of the synchronization code internally treats deadlines which are equivalent to `time_point::max()` for the same clock as a sentinel value indicating a deadline at infinity, or equivalently the lack of a deadline.

Care must be taken when converting between clocks to translate a deadline at infinity for one clock to the deadline at infinity for the other clock.

Reviewed By: nbronson

Differential Revision: D6720021

fbshipit-source-id: cfc230dd2d8db55297385a4afcb6d87ae4221840