fix building with CMake 3.9+
Summary:
D6806326 reduced the minimum version of CMake required to 3.2.0.  This had the
unintended effect of also disabling some newer CMake policies by default, which
caused builds to fail in newer CMake versions.

In particular, we make use of the GoogleTest module when it is available (CMake
3.9+).  However it requires policy `CMP0057` to be enabled.

This explicitly enables `CMP0057` when it is available.

Reviewed By: yfeldblum

Differential Revision: D6816891

fbshipit-source-id: 60ca46654d46207a96f0fad9c3467426a25b4afd