fix SingletonTestGlobal in shared library builds
Summary:
The SingletonTestGlobal code asserted that there was exactly one singleton
object registered (the one defined in this test file).  However if folly is
built as a shared library, all of it will be loaded, and several other
singletons are also defined by other parts of folly.  (Currently 7 singletons
get registered.)

Reviewed By: yfeldblum

Differential Revision: D6817544

fbshipit-source-id: de3947770690ffb68d097040bbbfbe42bce3bf94