Add missing override for adjacent_tokens_only
Summary:
The UNIX builds treat all warnings as errors, and this missing `override` keyword causes a warning which is then a hard error in the build. In Boost 1.59 and later, `Boost.ProgramOptions` provides a virtual method `adjacent_tokens_only()` which the class `GFlagValueSemanticBase` wants to override. As such, mark it using `override`.
Closes https://github.com/facebook/folly/pull/743

Differential Revision: D6851455

Pulled By: yfeldblum

fbshipit-source-id: 5ad448f0aad8f51f4328bfe9af37d63023089e88