fix buffer math in AsyncSocketTest.WriteErrorCallbackBytesWritten
Summary:
The logic in `AsyncSocketTest.WriteErrorCallbackBytesWritten` relied on overly
strict assumptions about how socket buffering would be implemented by the
kernel.  The tests would fail spuriously if the kernel allowed more data to be
buffered than the test assumed.

This updates the tests to send more data and allow the kernel to buffer more
data than it initially expects.

Reviewed By: yfeldblum

Differential Revision: D6843330

fbshipit-source-id: a53e19028d50c5d646654e67518a80d6b68106a2